ajout de l api key

+++ src/v1/bill.php
@@ -4,12 +4,15 @@ use \Psr\Http\Message\ServerRequestInterface as Request;
 $app->group('/bill', function () use ($app) {
     $app->post('/add', function (Request $request, Response $response, $args) {
         $data = $request->getParsedBody();
+        $api = new Api();
+        if (!$api->checkout($data['api_key'])){
+            return $response->withJson(['success' => false]);
+        }
         $dev = new Bills();
         $dev->date_start = $data['date_start'];
         $dev->date_end = $data['date_end'];
         $dev->price = $data['price'];
-        $dev->user_id = $data['user_id'];
-        $dev->offer_id = $data['offer_id'];
+        $dev->idcommands = $data['idcommands'];
         $dev->save();
         return $response->withJson(['success' => true]);
     });
@@ -20,12 +23,20 @@ $app->group('/bill', function () use ($app) {
 
     $app->post('/delete', function (Request $request, Response $response, $args) {
         $data = $request->getParsedBody();
+        $api = new Api();
+        if (!$api->checkout($data['api_key'])){
+            return $response->withJson(['success' => false]);
+        }
         Bills::where('idfacture', $data['id'])->delete();
         return $response->withJson(['success' => true]);
     });
 
     $app->post('/update', function (Request $request, Response $response, $args) {
         $data = $request->getParsedBody();
+        $api = new Api();
+        if (!$api->checkout($data['api_key'])){
+            return $response->withJson(['success' => false]);
+        }
         foreach ($data['data'] as $key => $item){
             Bills::where('idfacture', $data['id'])
                 ->update([$key => $item]);
